refactor: 빌드 에러 해결결
