+++ package.json
@@ -20,6 +20,7 @@
     "@tabler/icons-react": "^3.26.0",
     "@tanstack/react-query": "^5.62.11",
     "@tanstack/react-query-devtools": "^5.62.11",
+    "@tanstack/react-query-next-experimental": "^5.69.0",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
