+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,91 +1,5 @@
 import NextAuth from 'next-auth';
-import type { NextAuthOptions } from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: { label: '이메일', type: 'email' },
-        password: { label: '비밀번호', type: 'password' },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) return null;
-
-        try {
-          const res = await fetch(
-            `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/member/login`,
-            {
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/json',
-              },
-              body: JSON.stringify({
-                email: credentials.email,
-                password: credentials.password,
-              }),
-            },
-          );
-
-          const data = await res.json();
-          const authHeader = res.headers.get('authorization') || undefined;
-          const refreshToken = res.headers.get('refresh-token') || undefined;
-
-          console.log('로그인 응답 데이터:', data);
-          console.log('인증 헤더:', authHeader);
-
-          if (data.status === 200 && data.data) {
-            const user = {
-              id: data.data.email,
-              email: data.data.email,
-              nickname: data.data.nickname,
-              role: data.data.role,
-              accessToken: authHeader,
-              refreshToken: refreshToken,
-            };
-            console.log('생성된 사용자 객체:', user);
-            return user;
-          }
-          return null;
-        } catch (error) {
-          console.error('Auth error:', error);
-          return null;
-        }
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.email = user.email;
-        token.nickname = user.nickname;
-        token.role = user.role;
-        token.accessToken = user.accessToken;
-        token.refreshToken = user.refreshToken;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.email = token.email;
-        session.user.nickname = token.nickname;
-        session.user.role = token.role;
-        session.user.accessToken = token.accessToken;
-        session.user.refreshToken = token.refreshToken;
-      }
-      return session;
-    },
-  },
-  pages: {
-    signIn: '/signin/email',
-  },
-  session: {
-    strategy: 'jwt',
-    maxAge: 60 * 60, // 1 hour
-  },
-  secret: process.env.NEXTAUTH_SECRET,
-};
+import { authOptions } from './options';
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
