+++ src/components/ui/tooltip.tsx
@@ -1,30 +1,32 @@
-'use client';
+// 에러 처리를 위해 가림 처리함. 사용하지 않을 경우 삭제합니다.
 
-import * as React from 'react';
-import * as TooltipPrimitive from '@radix-ui/react-tooltip';
+// 'use client';
 
-import { cn } from '@/lib/utils';
+// import * as React from 'react';
+// import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 
-const TooltipProvider = TooltipPrimitive.Provider;
+// import { cn } from '@/lib/utils';
 
-const Tooltip = TooltipPrimitive.Root;
+// const TooltipProvider = TooltipPrimitive.Provider;
 
-const TooltipTrigger = TooltipPrimitive.Trigger;
+// const Tooltip = TooltipPrimitive.Root;
 
-const TooltipContent = React.forwardRef<
-  React.ElementRef<typeof TooltipPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof TooltipPrimitive.Content>
->(({ className, sideOffset = 4, ...props }, ref) => (
-  <TooltipPrimitive.Content
-    ref={ref}
-    sideOffset={sideOffset}
-    className={cn(
-      'z-50 overflow-hidden rounded-md border border-neutral-200 bg-white px-3 py-1.5 text-sm text-neutral-950 shadow-md animate-in fade-in-0 zoom-in-95 data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=closed]:zoom-out-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 dark:border-neutral-800 dark:bg-neutral-950 dark:text-neutral-50',
-      className,
-    )}
-    {...props}
-  />
-));
-TooltipContent.displayName = TooltipPrimitive.Content.displayName;
+// const TooltipTrigger = TooltipPrimitive.Trigger;
 
-export { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider };
+// const TooltipContent = React.forwardRef<
+//   React.ElementRef<typeof TooltipPrimitive.Content>,
+//   React.ComponentPropsWithoutRef<typeof TooltipPrimitive.Content>
+// >(({ className, sideOffset = 4, ...props }, ref) => (
+//   <TooltipPrimitive.Content
+//     ref={ref}
+//     sideOffset={sideOffset}
+//     className={cn(
+//       'z-50 overflow-hidden rounded-md border border-neutral-200 bg-white px-3 py-1.5 text-sm text-neutral-950 shadow-md animate-in fade-in-0 zoom-in-95 data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=closed]:zoom-out-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 dark:border-neutral-800 dark:bg-neutral-950 dark:text-neutral-50',
+//       className,
+//     )}
+//     {...props}
+//   />
+// ));
+// TooltipContent.displayName = TooltipPrimitive.Content.displayName;
+
+// export { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider };
