--- src/providers/RecoilRootProvider.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import { RecoilRoot } from 'recoil';
-
-const RecoilRootProvider = ({ children }: { children: React.ReactNode }) => {
-  return <RecoilRoot>{children}</RecoilRoot>;
-};
-
-export default RecoilRootProvider;
